refactor(digest): move the routes into their own router
Two reasons, one architectural and one practical.

`analysis.py` describes itself as clash detection and IDS validation. A model
digest is neither, and appending it there made a 1300-line module longer for no
reason other than that the file was already open.

The practical one: concurrent work is actively editing `analysis.py`. The two
changes were ~700 lines apart and merged clean, but "clean today" is not a
property worth relying on when the alternative is not sharing the file at all.
This branch no longer touches `analysis.py`; `git diff main -- analysis.py` is
empty.

No behaviour change — same two paths, same handlers. digest_route (25 checks)
re-verified green; ruff clean; route-authz, reachable and route-order pass.
